% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_pequod_c}
\alias{scale_color_pequod_c}
\alias{scale_colour_pequod_c}
\alias{scale_fill_pequod_c}
\title{Pequod continuous colour / fill scales for ggplot2}
\usage{
scale_color_pequod_c(palette = "log", reverse = FALSE, ...)

scale_colour_pequod_c(palette = "log", reverse = FALSE, ...)

scale_fill_pequod_c(palette = "log", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Palette name. See [palette_pequod()].}

\item{reverse}{Reverse the palette order.}

\item{...}{Further arguments passed to
[ggplot2::scale_color_gradientn()] / [ggplot2::scale_fill_gradientn()].}
}
\value{
A ggplot2 scale.
}
\description{
Interpolates across a Pequod palette with
[ggplot2::scale_color_gradientn()]. Best used with the `"log"` or
`"log-cool"` palettes, which are ordered from light to dark; the
`"crew"` palettes are categorical and will not interpolate cleanly.
}
